Move to manage cases directory and add correct test tags
