+++ e2e/journeys/manageCases/caseProgression/checkApplication/checkApplication.ts
@@ -1,7 +1,7 @@
 import { Page } from "@playwright/test";
-import { Fl401AddCaseNumber1Page } from "../../../../pages/citizen/caseProgression/checkApplication/fl401AddCaseNumber1Page";
+import { Fl401AddCaseNumber1Page } from "../../../../pages/manageCases/caseProgression/checkApplication/fl401AddCaseNumber1Page";
 import { Helpers } from "../../../../common/helpers";
-import { Fl401AddCaseNumberSubmitPage } from "../../../../pages/citizen/caseProgression/checkApplication/fl401AddCaseNumberSubmitPage";
+import { Fl401AddCaseNumberSubmitPage } from "../../../../pages/manageCases/caseProgression/checkApplication/fl401AddCaseNumberSubmitPage";
 
 interface CheckApplicationParams {
   page: Page;
