+++ e2e/tests/manageCases/caseProgression/checkApplication/checkApplication.spec.ts
@@ -1,19 +1,19 @@
 import { test } from "@playwright/test";
 import Config from "../../../../config";
 import createDaCitizenCourtNavCase from "../../../../common/createCaseHelper";
-import { CheckApplication } from "../../../../journeys/citizen/caseProgression/checkApplication/checkApplication";
+import { CheckApplication } from "../../../../journeys/manageCases/caseProgression/checkApplication/checkApplication";
 import { Helpers } from "../../../../common/helpers";
 import config from "../../../../config";
 
 test.use({ storageState: Config.sessionStoragePath + "caseWorker.json" });
 
-test.describe("Complete Check Application task for DA Citizen case. @citizenFrontend @crossbrowserCitizenFrontend", () => {
+test.describe("Check Application task for DA Citizen case tests. @manageCases", () => {
   test.beforeEach(async ({ page }) => {
     const ccdRef: string = await createDaCitizenCourtNavCase(false);
     await Helpers.goToCase(page, config.manageCasesBaseURL, ccdRef, "tasks");
   });
 
-  test("Complete Check Application without accessibility test.", async ({
+  test("Complete Check Application without accessibility test. @crossbrowserManageCases", async ({
     page,
   }): Promise<void> => {
     await CheckApplication.checkApplication({
@@ -21,15 +21,8 @@ test.describe("Complete Check Application task for DA Citizen case. @citizenFron
       accessibilityTest: false,
     });
   });
-});
 
-test.describe("Testing the accessibility when completing Check Application task for DA Citizen case. @accessibilityCitizenFrontend", () => {
-  test.beforeEach(async ({ page }) => {
-    const ccdRef: string = await createDaCitizenCourtNavCase(false);
-    await Helpers.goToCase(page, config.manageCasesBaseURL, ccdRef, "tasks");
-  });
-
-  test("Complete Check Application with accessibility test.", async ({
+  test("Complete Check Application with accessibility test. @accessibilityManageCases", async ({
     page,
   }): Promise<void> => {
     await CheckApplication.checkApplication({
